feat: add multi-plate slicing and service
