+++ src/swagger-ui-express.d.ts
@@ -1,0 +1,1 @@
+declare module 'swagger-ui-express';
